requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.0
sqlalchemy>=2.0.0
httpx>=0.23.0
beautifulsoup4>=4.11.0
apscheduler>=3.10.0
python-dotenv>=1.0.0
openai>=0.27.0
pydantic>=2.0.0
alembic>=1.10.0
requests>=2.28.0 